Port zur neusten version komplett (1.21.5)

+++ src/main/java/mett/palemannie/tabakmod/block/custom/DunklerTabakBallenBlock.java
@@ -27,7 +27,7 @@ public class DunklerTabakBallenBlock extends Block {
     }
 
     @Override
-    public void fallOn(Level level, BlockState blockState, BlockPos blockPos, Entity entity, float v) {
+    public void fallOn(Level level, BlockState blockState, BlockPos blockPos, Entity entity, double v) {
         entity.causeFallDamage(v, 0.45f, level.damageSources().fall());
     }
     //------------------------------------------------------------------------------------------------------------------
